test: move spec file and setup basic struct
